+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - 增加命令行功能 
+- 优化消费器错误信息
+- 解决github.com/siddontang/go-mysql/client线程不安全问题
 
 ## [1.1.0] 2019-12-09
 ### Added